requirements.txt
pdfplumber
simplejson
requests
PyYAML
pillow
reportlab
pandas
loguru
openai
Flask